requirements.txt
flask>=2.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.9
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
fastmcp>=0.1.0
beautifulsoup4>=4.12.2
httpx>=0.24.0
mcp>=0.1.0
